Live/requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
psycopg2-binary>=2.9.0
aiohttp>=3.9.0
requests>=2.31.0
